Surface timezone editor on the Profile tab
