+++ frontend/public/locales/en/settings.json
@@ -48,6 +48,10 @@
     "newPasswordLabel": "New password",
     "confirmPasswordLabel": "Confirm password",
     "passwordPlaceholder": "••••••••",
+    "timezoneLabel": "Timezone",
+    "timezonePlaceholder": "Select timezone",
+    "timezoneEmpty": "No timezone found.",
+    "timezoneHelp": "Used for due dates, daily reminders, and recurring task math. Detected from your browser when you registered; change it here if it's wrong.",
     "avatarLabel": "Avatar",
     "avatarUploadTab": "Upload",
     "avatarUrlTab": "URL",
