+++ frontend/public/locales/es/settings.json
@@ -48,6 +48,10 @@
     "newPasswordLabel": "Nueva contraseña",
     "confirmPasswordLabel": "Confirmar contraseña",
     "passwordPlaceholder": "••••••••",
+    "timezoneLabel": "Zona horaria",
+    "timezonePlaceholder": "Selecciona una zona horaria",
+    "timezoneEmpty": "No se encontró ninguna zona horaria.",
+    "timezoneHelp": "Se usa para las fechas de vencimiento, los recordatorios diarios y los cálculos de tareas recurrentes. Se detectó desde tu navegador al registrarte; cámbiala aquí si es incorrecta.",
     "avatarLabel": "Avatar",
     "avatarUploadTab": "Subir",
     "avatarUrlTab": "URL",
