+++ frontend/public/locales/fr/settings.json
@@ -48,6 +48,10 @@
     "newPasswordLabel": "Nouveau mot de passe",
     "confirmPasswordLabel": "Confirmer le mot de passe",
     "passwordPlaceholder": "••••••••",
+    "timezoneLabel": "Fuseau horaire",
+    "timezonePlaceholder": "Sélectionner un fuseau horaire",
+    "timezoneEmpty": "Aucun fuseau horaire trouvé.",
+    "timezoneHelp": "Utilisé pour les échéances, les rappels quotidiens et les calculs de tâches récurrentes. Détecté depuis votre navigateur lors de l'inscription ; modifiez-le ici s'il est incorrect.",
     "avatarLabel": "Avatar",
     "avatarUploadTab": "Télécharger",
     "avatarUrlTab": "URL",
